trail primitive 수정중

+++ iitp-rest/src/main/java/com/iitp/iitp_rest/model/publicTransit/rail/ExitXml.java
@@ -16,7 +16,7 @@ public class ExitXml {
     @XmlAttribute
     private Double offset;
     @XmlAttribute
-    private Long accessTime;
+    private Double accessTime;
     @XmlAttribute
     private String coord;
 }
